FIX:로그인기능 / FEAT:회원가입 기능 디테일화
