+++ src/views/examples/ResultPage.jsx
